requirements.txt
# --------- pytorch --------- #
timm
# torch>=1.10.0
# torchvision>=0.11.0
pytorch-lightning
torchmetrics==0.9.3
fairscale==0.4.12

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# demo

gradio==3.3.1

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)

# --------- others --------- #
alibi-detect[torch]
albumentations
torchdrift